+++ native/reg_command.h
@@ -1,7 +1,7 @@
 #ifndef REG_COMMAND_H
 #define REG_COMMAND_H
 
-#include <windows.h>
+#include <windef.h>
 
 #include "macros.h"
 
@@ -10,7 +10,7 @@ void do_write_reg_command(FILE *out_fp,
                           const wchar_t *prop,
                           const char *value,
                           size_t value_len,
-                          DWORD type,
+                          unsigned long type,
                           bool debug);
 void do_write_reg_commands(
     FILE *out_fp, HKEY hk, unsigned n_values, const wchar_t *prior_stem, bool debug);
